pyproject.toml
[tool.poetry]
name = "stability-matrix-tools"
version = "0.2.5"
description = ""
authors = ["Ionite <dev@ionite.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = {extras = ["dotenv"], version = "^2.4.2"}
semver = "^3.0.2"
httpx = "^0.25.0"
b2sdk = "^1.24.1"
python-dotenv = "^1.0.0"
cryptography = "^41.0.4"
keyring = "^24.2.0"
pyperclip = "^1.8.2"
blake3 = "^0.3.3"
pygithub = "^2.1.1"

[tool.poetry.scripts]
sm-tools = "stability_matrix_tools.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
